getEventsFromCategory endpoint implemented

+++ Awesome City Trips/Awesome City Trips/Models/Coordinates.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 struct Coordinates: Codable {
-    let longitude: Int64
-    let latitude: Int64
+    let longitude: Double
+    let latitude: Double
 }
